# packages/ts-sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cocoon-SIT API",
    "description": "Cocoon-SIT backend implementing REST 202 + WebSocket realtime contracts, versioned prompt templates, audit artifacts, and worker-driven runtime execution.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/cocoons": {
      "get": {
        "tags": [
          "cocoons"
        ],
        "summary": "List Cocoons",
        "operationId": "list_cocoons_api_v1_cocoons_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CocoonOut"
                  },
                  "type": "array",
                  "title": "Response List Cocoons Api V1 Cocoons Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "cocoons"
        ],
        "summary": "Create Cocoon",
        "operationId": "create_cocoon_api_v1_cocoons_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CocoonCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CocoonOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/cocoons/{cocoon_id}": {
      "patch": {
        "tags": [
          "cocoons"
        ],
        "summary": "Update Cocoon",
        "operationId": "update_cocoon_api_v1_cocoons__cocoon_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cocoon_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cocoon Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CocoonUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CocoonOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "cocoons"
        ],
        "summary": "Get Cocoon",
        "operationId": "get_cocoon_api_v1_cocoons__cocoon_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cocoon_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cocoon Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CocoonOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "cocoons"
        ],
        "summary": "Delete Cocoon",
        "operationId": "delete_cocoon_api_v1_cocoons__cocoon_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cocoon_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cocoon Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CocoonOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cocoons/tree": {
      "get": {
        "tags": [
          "cocoons"
        ],
        "summary": "Cocoon Tree",
        "operationId": "cocoon_tree_api_v1_cocoons_tree_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CocoonTreeNode"
                  },
                  "type": "array",
                  "title": "Response Cocoon Tree Api V1 Cocoons Tree Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/cocoons/{cocoon_id}/state": {
      "get": {
        "tags": [
          "cocoons"
        ],
        "summary": "Get Session State",
        "operationId": "get_session_state_api_v1_cocoons__cocoon_id__state_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cocoon_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cocoon Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStateOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CocoonCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "character_id": {
            "type": "string",
            "title": "Character Id"
          },
          "selected_model_id": {
            "type": "string",
            "title": "Selected Model Id"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "default_temperature": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Temperature"
          },
          "max_context_messages": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Context Messages"
          },
          "auto_compaction_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Compaction Enabled"
          }
        },
        "type": "object",
        "required": [
          "name",
          "character_id",
          "selected_model_id"
        ],
        "title": "CocoonCreate"
      },
      "CocoonOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "owner_user_id": {
            "type": "string",
            "title": "Owner User Id"
          },
          "character_id": {
            "type": "string",
            "title": "Character Id"
          },
          "selected_model_id": {
            "type": "string",
            "title": "Selected Model Id"
          },
          "default_temperature": {
            "type": "number",
            "title": "Default Temperature"
          },
          "max_context_messages": {
            "type": "integer",
            "title": "Max Context Messages"
          },
          "auto_compaction_enabled": {
            "type": "boolean",
            "title": "Auto Compaction Enabled"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "owner_user_id",
          "character_id",
          "selected_model_id",
          "default_temperature",
          "max_context_messages",
          "auto_compaction_enabled",
          "parent_id",
          "created_at"
        ],
        "title": "CocoonOut"
      },
...
      "CocoonTreeNode": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "children": {
            "items": {
              "$ref": "#/components/schemas/CocoonTreeNode"
            },
            "type": "array",
            "title": "Children"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "parent_id"
        ],
        "title": "CocoonTreeNode"
      },
      "CocoonUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "character_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Character Id"
          },
          "selected_model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Model Id"
          },
          "default_temperature": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Temperature"
          },
          "max_context_messages": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Context Messages"
          },
          "auto_compaction_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Compaction Enabled"
          }
        },
        "type": "object",
        "title": "CocoonUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SessionStateOut": {
        "properties": {
          "cocoon_id": {
            "type": "string",
            "title": "Cocoon Id"
          },
          "relation_score": {
            "type": "integer",
            "title": "Relation Score"
          },
          "persona_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Persona Json"
          },
          "active_tags_json": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Active Tags Json"
          },
          "current_wakeup_task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Wakeup Task Id"
          }
        },
        "type": "object",
        "required": [
          "cocoon_id",
          "relation_score",
          "persona_json",
          "active_tags_json",
          "current_wakeup_task_id"
        ],
        "title": "SessionStateOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}